pyproject.toml
[tool.poetry]
name = "admin"
version = "0.1.0"
description = ""
authors = ["Ivan Panchenko <arigatory@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-split-settings = "^1.3.1"
uwsgi = "^2.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
